gle: Merge all gamelogic engine interfaces into one.
