+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Game/CameraSettingInspector.cs
@@ -15,7 +15,6 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 using UnityEditor;
-using UnityEngine;
 
 namespace VisualPinball.Unity.Editor
 {
