+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Utils/Icons.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using UnityEditor;
-using UnityEditor.Callbacks;
 using UnityEngine;
 using VisualPinball.Unity.Playfield;
 using Object = UnityEngine.Object;
