--- VisualPinball.Unity/VisualPinball.Unity/Game/Engine/IGamelogicEngineWithCoils.cs
@@ -1,59 +1,0 @@
-﻿// Visual Pinball Engine
-// Copyright (C) 2021 freezy and VPE Team
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-using System;
-using VisualPinball.Engine.Game.Engines;
-
-namespace VisualPinball.Unity
-{
-	/// <summary>
-	/// A game logic engine working with coils (solenoids). <p/>
-	///
-	/// It provides a list of available coils and an event handler to trigger
-	/// them.
-	/// </summary>
-	public interface IGamelogicEngineWithCoils
-	{
-		/// <summary>
-		/// A list of available coils.
-		/// </summary>
-		GamelogicEngineCoil[] AvailableCoils { get; }
-
-		/// <summary>
-		/// Triggered when a coil is enabled or disabled.
-		/// </summary>
-		event EventHandler<CoilEventArgs> OnCoilChanged;
-	}
-
-	public readonly struct CoilEventArgs
-	{
-		/// <summary>
-		/// Id of the coil, as defined by <see cref="IGamelogicEngineWithCoils.AvailableCoils"/>.
-		/// </summary>
-		public readonly string Id;
-
-		/// <summary>
-		/// State of the coil, true if the coil is under voltage, false if not.
-		/// </summary>
-		public readonly bool IsEnabled;
-
-		public CoilEventArgs(string id, bool isEnabled)
-		{
-			Id = id;
-			IsEnabled = isEnabled;
-		}
-	}
-}
