--- VisualPinball.Unity/VisualPinball.Unity/Game/Engine/IGamelogicEngineWithCoils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c499fa5e8afc876429f3ef828771eae9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
