--- VisualPinball.Unity/VisualPinball.Unity/Game/Engine/IGamelogicEngineWithLamps.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 17c472a5cc0f74c4ba0d741649f50198
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
