--- VisualPinball.Unity/VisualPinball.Unity/Game/Engine/IGamelogicEngineWithSwitches.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d204e89151725944ae60b59c3bfcbac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
